microbot/requirements.txt
altair==4.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
base58==2.1.0
beautifulsoup4==4.9.3
black==21.9b0
bleach==3.3.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
cursor==1.3.4
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.6
entrypoints==0.3
filelock==3.0.12
gitdb==4.0.7
GitPython==3.1.18
google==3.0.0
googlesearch-python==1.0.1
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
platformdirs==2.3.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pychalk==2.0.1
pycparser==2.20
pydeck==0.6.2
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==301
pywinpty==1.1.2
pyzmq==22.1.0
regex==2021.9.24
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
smmap==4.0.0
soupsieve==2.2.1
sqlparse==0.4.1
streamlit==0.83.0
terminado==0.10.1
terminal-animation==0.6
testpath==0.5.0
toml==0.10.2
tomli==1.2.1
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
virtualenv==20.4.4
virtualenv-clone==0.5.4
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
